Add the form for new and edit
