+++ app/views/api/admin/seasons/index.json.jbuilder
@@ -1,6 +1,7 @@
 json.seasons do
   json.array! @seasons do |season|
     json.id season.id
+    json.anime_id season.anime_id
     json.phase season.phase
     json.name season.name
     json.start_on season.start_on
