+++ spec/requests/admin/seasons_spec.rb
@@ -14,6 +14,7 @@ describe 'GET /api/admin/animes/1/seasons', autodoc: true do
       seasons: [
         {
           id: season1.id,
+          anime_id: season1.anime_id,
           phase: season1.phase,
           name: season1.name,
           start_on: season1.start_on.strftime('%Y-%m-%d'),
@@ -22,6 +23,7 @@ describe 'GET /api/admin/animes/1/seasons', autodoc: true do
         },
         {
           id: season2.id,
+          anime_id: season2.anime_id,
           phase: season2.phase,
           name: season2.name,
           start_on: season2.start_on.strftime('%Y-%m-%d'),
